[implant] feat(stix): delete file after download on drop files (#3511)
